refactor(llc): group `validateAddress` by families
