+++ .changeset/mighty-houses-judge.md
@@ -1,0 +1,5 @@
+---
+"@ledgerhq/live-common": minor
+---
+
+refactor(llc): group validate address by families
